Fixed project.py complexity warning from code climate https://codeclimate.com/github/eviltnan/freeturn/pull/54
