+++ crm/wagtail_admin/project.py
@@ -12,23 +12,21 @@ from django.utils import timezone
 from django_filters.fields import ModelChoiceField
 from django_fsm import TransitionNotAllowed
 from google.api_core.exceptions import GoogleAPIError
-from google.auth.exceptions import GoogleAuthError
 from instance_selector.widgets import InstanceSelectorWidget
 from wagtail.admin import messages
 from wagtail.admin.forms import WagtailAdminModelForm
 from wagtail.admin.rich_text import get_rich_text_editor_widget
 from wagtail.admin.search import SearchArea
-from wagtail.contrib.modeladmin.helpers import AdminURLHelper, ButtonHelper, PermissionHelper
+from wagtail.contrib.modeladmin.helpers import AdminURLHelper, ButtonHelper
 from wagtail.contrib.modeladmin.mixins import ThumbnailMixin
 from wagtail.contrib.modeladmin.options import ModelAdmin
-from wagtail.contrib.modeladmin.views import CreateView, InspectView, IndexView, ModelFormView, \
+from wagtail.contrib.modeladmin.views import CreateView, InspectView, ModelFormView, \
     InstanceSpecificView
 from wagtail.tests.utils.form_data import rich_text
 
 from crm import gmail_utils
 from crm.models import City, CV, MessageTemplate
 from crm.models.project import Project
-from crm.models.project_message import ProjectMessage
 
 logger = logging.getLogger(__file__)
 
@@ -285,43 +283,3 @@ class ProjectSearchArea(SearchArea):
             name='projects',
             classnames='icon icon-fa-product-hunt',
             order=101)
-
-
-class MessagePermissionHelper(PermissionHelper):
-    def user_can_create(self, user):
-        return False
-
-
-class ProjectMessageIndexView(IndexView):
-    def get_context_data(self, **kwargs):
-        try:
-            created_messages = gmail_utils.sync()
-        except GoogleAuthError as ex:
-            logger.error(f"Can't update messages: {ex}")
-            messages.error(self.request, f"Can't update messages: {ex}")
-            created_messages = []
-        if created_messages:
-            messages.info(
-                self.request,
-                f'{len(created_messages)} new messages'
-            )
-        return super().get_context_data(**kwargs)
-
-
-class MessageAdmin(ModelAdmin):
-    model = ProjectMessage
-    menu_icon = 'fa-envelope-open'
-    menu_label = 'Messages'
-    list_display = ['subject', 'author', 'project', 'created']
-    list_filter = ['project', 'author']
-    ordering = ['-created']
-    index_view_class = ProjectMessageIndexView
-    inspect_view_enabled = True
-    inspect_view_fields = ['project', 'subject', 'author', 'text']
-    inspect_template_name = 'message_inspect.html'
-    permission_helper_class = MessagePermissionHelper
-    search_fields = ['subject',
-                     'author__first_name',
-                     'author__last_name',
-                     'project__name',
-                     'project__company__name']
